requirements.txt
fireworks-ai==0.15.2
nomic==3.0.15
openai==1.30.1
pymongo==4.7.2
sentence-transformers==2.2.2
requests==2.31.0
mistral==18.0.1
ipykernel==6.29.5
